+++ app/images/static/css/obs-styles.css
@@ -327,7 +327,6 @@ div.searchbox .searchbtn { padding: 3px 8px 3px 8px; }
 div.searchboxbig .searchbtn { padding: 5px 8px 5px 8px; }
 
 img.observation-image {
-	width: 560px;
 	z-index: 1;
 }
