+++ app/images/templates/images/archive_image.html
@@ -3,6 +3,26 @@
 {% load staticfiles %}
 {% block header_script %}
 <script type="text/javascript" src="{% static "js/rti_archive_observation.js" %}"></script>
+<script type="text/javascript">
+$(document).ready(function(){
+
+var frame_id = {{frameid}}
+var img_url;
+
+  $.get("https://thumbnails.lcogt.net/"+ frame_id+"/?height=1000&width=1000", function(data){
+    img_url = data.url;
+    $('img.observation-image').attr('src',img_url);
+    $('a.observation-image-link').attr('href', img_url)
+  });
+
+  $.get("https://archive-api.lcogt.net/frames/"+ frame_id+"/", function(data){
+    var download_content = "<a href='"+ data.url+"' property='UCD:obs.image'>FITS</a> or <a href='"+ img_url+"' property='UCD:obs.image'>JPEG</a>";
+    $('#download-link').html(download_content);
+  });
+
+});
+
+</script>
 {% endblock %}
 
 {%block header %}{{ obs.OBJECT }}{%endblock%}
@@ -19,8 +39,8 @@
 {% endblock %}
 
 {% block main-content %}
-        <a href='{{ obs.fullimage_url }}' title='Click for full size version'>
-            <img src='{{ obs.fullimage_url }}' class='observation-image' alt='{{ obs.title }}' />
+        <a href='{{ obs.fullimage_url }}' title='Click for full size version' class="observation-image-link">
+            <img src='{{ obs.fullimage_url }}' class='observation-image img-responsive' alt='{{ obs.title }}' />
         </a>
         {% if otherobs%}
         <p class='moreobservations'>More observations by {{ obs.user }}:</p>
@@ -28,7 +48,7 @@
             {% for ob in otherobs %}
             <li class='{{ ob.class }}'>
                 <a href='{{ ob.url }}'>
-                    <img src='{{ ob.thumb }}' alt='{{ ob.title }}' title='{{ ob.title }} taken on {{ ob.date }}' />
+                    <img src='{{ ob.thumb }}' alt='{{ ob.title }}' title='{{ ob.title }} taken on {{ ob.date }}'/>
                 </a>
             </li>
             {% endfor %}
@@ -48,25 +68,25 @@
                 <tr>
                     <th scope="row">Observer:</th>
                     <td>
-                        <span property="UCD:obs.observer">{{ obs.user }}</span>
+                        <span property="UCD:obs.observer">{{ observer }}</span>
                     </td>
                 </tr>
                 <tr>
                     <th scope="row">Date:</th>
                     <td>
-                        <time property='UCD:time.creation'>{{ obs.dateobs|date:"c" }}</time>
+                        <time property='UCD:time.creation'>{{ obs.DATE_OBS }}</time>
                     </td>
                 </tr>
                 <tr>
                     <th scope="row">Coordinates:</th>
                     <td>
-                        <span class='Spatial-ReferenceValue' title='{{ obs.ra }};{{ obs.dec }}'>
-                            <span property='UCD:pos.eq.RA' content='{{ obs.ra }}'>{{ obs.ra|degreestohms }}</span>,
-                            <span property='UCD:pos.eq.DEC' content='{{ obs.dec }}'>{{ obs.dec|degreestodms }}</span>
+                        <span class='Spatial-ReferenceValue' title='{{ obs.RA }};{{ obs.DEC }}'>
+                            <span property='UCD:pos.eq.RA' content='{{ obs.RA|hmstodegrees }}'>{{ obs.RA }}</span>,
+                            <span property='UCD:pos.eq.DEC' content='{{ obs.DEC|dmstodegrees }}'>{{ obs.DEC }}</span>
                         </span>
                         <span class='Spatial-CoordinateFrame'>FK5</span>
                         <span class='Spatial-Equinox'>J2000</span>
-                        <br />View this position in <a href='http://server1.wikisky.org/v2?ra={{ obs.ra|hmstohours }}&amp;de={{ obs.dec|dmstodegrees }}&amp;zoom=10&amp;img_source=astrophoto' title="View these coordinates in Wikisky">Wikisky</a> or <a href='http://www.worldwidetelescope.org/wwtweb/goto.aspx?object=ViewShortcut&amp;ra={{ obs.ra|hmstohours }}&amp;dec={{ obs.dec|dmstodegrees }}&amp;zoom=3' title="View these coordinates in Microsoft's World Wide Telescope">WorldWideTelescope</a>.
+                        <br />View this position in <a href='http://server1.wikisky.org/v2?ra={{ obs.RA|hmstohours }}&amp;de={{ obs.DEC|dmstodegrees }}&amp;zoom=10&amp;img_source=astrophoto' title="View these coordinates in Wikisky">Wikisky</a> or <a href='http://www.worldwidetelescope.org/wwtweb/goto.aspx?object=ViewShortcut&amp;ra={{ obs.RA|hmstohours }}&amp;dec={{ obs.DEC|dmstodegrees }}&amp;zoom=3' title="View these coordinates in Microsoft's World Wide Telescope">WorldWideTelescope</a>.
                     </td>
                 </tr>
                 <tr>
@@ -86,18 +106,9 @@
                 <tr>
                     <th scope="row">Data:</th>
                     <td>
-                        {% if filters %}
-                        <ul class='filters'>
-                            {% for filter in filters %} {% if filter.fits %}
-                            <li class='filter-{{ filter.id }}'><a href='{{ filter.fits }}' property='UCD:obs.image'>{% if forloop.first  and forloop.last %}{% else %}{{ filter.fullname }} {%endif%}FITS</a> or <a href='{{ filter.img }}' title='{{ filter.name }} jpeg' class='filter-{{ filter.name }}'>jpeg</a>
-                            </li>
-                            {% endif %} {% endfor %}
-                        </ul>
-                        {% else %} No FITS data found
-                        <br />{% endif %}
-        {% if not framedb_obs %}
-                        <a href='http://www.worldwidetelescope.org/wwtweb/ShowImage.aspx?name={{ obs.object }}&ra={{ obs.ra }}&dec={{ obs.dec }}8&x=1024&y=1024&scale=0.274&rotation=0&imageurl=http://rti.images-telescope.com/observations/2011/08/09/process-677-1.jpg&creditsUrl=http://lcogt.net/observations/ogg/2m0a/61407&credits=Image%20taken%20with%20Faulkes%20Telescope%20North%20operated%20by%20Las%20Cumbres%20Observatory%20Global%20Telescope%20Network&thumb=http://rti.images-telescope.com/observations/2011/08/09/process-677-1_120.jpg'>View in WorldWideTelescope</a>
-        {% endif %}
+
+                        <div id="download-link">No FITS data found</div>
+                        <p><a href="http://lcogt.net/education/observing/fpack/">Having trouble opening FITS files?</a></p>
                     </td>
                 </tr>
                 <tr>
